Create inbuilt undo history
Disable pushing to browser history (which was inherited from
paloma and seemingly broken) and instead introduce an internal
history, and put an undo button into the menu bar.

This commit does not deal with the case where a 'typing'
notification appears when the undo button is pressed. In that
case, messages will be repeated. Waiting for a future commit.